Move some shared stuff from jws::compact into jws
This should not be externally visible.
